dom/workers/test/test_navigator.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Navigator
-->
<head>
  <title>Test for DOM Worker Navigator</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script>
  ok(!self.isSecureContext, "This test should not be running in a secure context");
</script>
<script type="text/javascript" src="test_navigator.js"></script>
</pre>
</body>
</html>
